Extract gwilym_encoding into its own file
